Remove some old references from the documentation
I'm not sure what this is trying to say. You can't **call** a
function with an optional parameter like the one in the example.

I did a test where I **defined** a function with an optional
parameter:

    function null_test($a = null) {
        if ($a === null) {
            return "Yes NULL!";
        } else {
            return "Not NULL!";
        }
    }

When I call this as `$client->null_test()` it reacts like I would
expectit to. When I call it as `$client->null_test("hello")` it
still responds like I would expect it to.

Perhaps this is something that's fixed in more modern PHPs? I tested
on PHP 7.1.